pyproject.toml
[project]
name = "zev"
version = "0.7.0"
description = "Lookup CLI commands easily"
readme = "README.md"
dependencies = [
    "google-genai>=1.11.0",
    "openai>=1.72.0",
    "pydantic>=2.10.6",
    "pyperclip>=1.9.0",
    "python-dotenv>=1.0.1",
    "questionary>=2.1.0",
    "rich>=13.9.4"
]
requires-python = ">=3.9"
urls = { Repository = "https://github.com/dtnewman/zev" }

[project.scripts]
zev = "zev.main:app"

[project.optional-dependencies]
dev = [
    "ruff>=0.11.2",
    "isort>=5.13.2"
]
azure = [
    "azure-identity>=1.20.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["zev*"]

[tool.ruff]
line-length = 120
# this ensures imports are properly sorted
lint.select = ["I001"]

[tool.isort]
profile = "black"
line_length = 120
